+++ safeincave/BC/Heat/handler.py
@@ -1,0 +1,263 @@
+# Copyright (c) 2026, The SafeInCave Developers
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+from __future__ import annotations
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from ...Cavern import CavernHandler
+
+from ..base import GeneralBC
+from ..cavern_utils import compute_cavern_heat_flux
+import numpy as np
+import dolfinx as do
+
+
+class BcHandler:
+    """
+    Boundary-condition handler for a heat-diffusion equation.
+
+    This class stores user-defined boundary conditions, organizes them by
+    type (Dirichlet, Neumann, Robin), and converts them into DOLFINx/UFL
+    objects at a given time ``t`` for assembly.
+
+    Parameters
+    ----------
+    equation : HeatDiffusion
+        Equation object providing:
+
+        - ``V``: function space of the primary field.
+        - ``grid``: mesh/grid handler with methods
+          :meth:`get_boundary_tags` and :meth:`get_boundary_tag`.
+        - Variational symbols, e.g., ``T_`` (test function), ``dT`` (trial
+          increment), and the measure ``ds``.
+
+    Attributes
+    ----------
+    eq : HeatDiffusion
+        Stored equation reference.
+    dirichlet_boundaries : list[DirichletBC]
+        Registered Dirichlet BCs.
+    neumann_boundaries : list[NeumannBC]
+        Registered Neumann BCs.
+    robin_boundaries : list[RobinBC]
+        Registered Robin BCs.
+    dirichlet_bcs : list[dolfinx.fem.DirichletBC]
+        DOLFINx Dirichlet BC objects at the current time (set by :meth:`update_dirichlet`).
+    neumann_bcs : list[ufl.Form]
+        Neumann contributions to the linear form (set by :meth:`update_neumann`).
+    robin_bcs_a : list[ufl.Form]
+        Robin contributions to the bilinear form (coefficient times trial/test; set by :meth:`update_robin`).
+    robin_bcs_b : list[ufl.Form]
+        Robin contributions to the linear form (coefficient times ambient value; set by :meth:`update_robin`).
+
+    Notes
+    -----
+    Time dependence is handled via piecewise-linear interpolation using
+    :func:`numpy.interp` between ``values`` and ``time_values`` stored in each
+    boundary-condition object.
+    """
+
+    def __init__(self):
+        self.dirichlet_boundaries = []
+        self.neumann_boundaries = []
+        self.robin_boundaries = []
+
+    def set_V(self, V):
+        self.V = V
+
+    def set_boudary_tags(self, boundary_tags):
+        self.boundary_tags = boundary_tags
+
+    def set_dolfin_tags(self, dolfin_tags):
+        self.dolfin_tags = dolfin_tags
+
+    def set_boundary_dim(self, boundary_dim):
+        self.boundary_dim = boundary_dim
+
+    def set_ds(self, ds):
+        self.ds = ds
+
+    def set_test_function(self, T_):
+        self.T_ = T_
+
+    def set_trial_function(self, dT):
+        self.dT = dT
+
+    def reset_boundary_conditions(self) -> None:
+        """
+        Clear all registered boundary conditions of all types.
+
+        Returns
+        -------
+        None
+        """
+        self.dirichlet_boundaries = []
+        self.neumann_boundaries = []
+        self.robin_boundaries = []
+
+    def add_boundary_condition(self, bc: GeneralBC) -> None:
+        """
+        Register a boundary condition by its type.
+
+        Parameters
+        ----------
+        bc : GeneralBC
+            One of :class:`DirichletBC`, :class:`NeumannBC`, or :class:`RobinBC`.
+
+        Returns
+        -------
+        None
+
+        Raises
+        ------
+        Exception
+            If the boundary condition type is not supported.
+        """
+        if bc.type == "dirichlet":
+            self.dirichlet_boundaries.append(bc)
+        elif bc.type == "neumann":
+            self.neumann_boundaries.append(bc)
+        elif bc.type == "robin":
+            self.robin_boundaries.append(bc)
+        else:
+            raise Exception(f"Boundary type {bc.type} not supported.")
+
+    def update_bcs(self, t: float) -> None:
+        """
+        Update all boundary-condition objects for a given time.
+
+        This builds the runtime lists:
+        ``dirichlet_bcs``, ``neumann_bcs``, ``robin_bcs_a``, and ``robin_bcs_b``.
+
+        Parameters
+        ----------
+        t : float
+            Current simulation time.
+
+        Returns
+        -------
+        None
+        """
+        self.update_dirichlet(t)
+        self.update_neumann(t)
+        self.update_robin(t)
+
+    def update_dirichlet(self, t: float) -> None:
+        """
+        Build Dirichlet BC objects at time ``t``.
+
+        Parameters
+        ----------
+        t : float
+            Current simulation time.
+
+        Returns
+        -------
+        None
+
+        Side Effects
+        ------------
+        Populates :attr:`dirichlet_bcs` with
+        :class:`dolfinx.fem.DirichletBC` objects constructed from:
+        - located dofs on the boundary,
+        - interpolated Dirichlet value at time ``t``.
+        """
+        self.dirichlet_bcs = []
+        for bc in self.dirichlet_boundaries:
+            value = np.interp(t, bc.time_values, bc.values)
+            dofs = do.fem.locate_dofs_topological(
+                self.V, self.boundary_dim, self.boundary_tags[bc.boundary_name]
+            )
+            self.dirichlet_bcs.append(
+                do.fem.dirichletbc(do.default_scalar_type(value), dofs, self.V)
+            )
+
+    def update_neumann(self, t: float) -> None:
+        """
+        Build Neumann contributions to the linear form at time ``t``.
+
+        Parameters
+        ----------
+        t : float
+            Current simulation time.
+
+        Returns
+        -------
+        None
+
+        Side Effects
+        ------------
+        Populates :attr:`neumann_bcs` with UFL terms of the form
+        ``value * eq.T_ * ds(tag)`` to be added to the right-hand side.
+        """
+        self.neumann_bcs = []
+        for bc in self.neumann_boundaries:
+            value = np.interp(t, bc.time_values, bc.values)
+            self.neumann_bcs.append(
+                value
+                * self.T_
+                * self.ds(self.dolfin_tags[self.boundary_dim][bc.boundary_name])
+            )
+
+    def update_robin(self, t: float) -> None:
+        """
+        Build Robin (convective) contributions to bilinear/linear forms.
+
+        Parameters
+        ----------
+        t : float
+            Current simulation time.
+
+        Returns
+        -------
+        None
+
+        Side Effects
+        ------------
+        Populates:
+        - :attr:`robin_bcs_a` with terms ``h * dT * T_ * ds(tag)``
+          (bilinear form contribution),
+        - :attr:`robin_bcs_b` with terms ``h * T_inf * T_ * ds(tag)``
+          (linear form contribution),
+        where ``T_inf`` is interpolated from the Robin BC values at time ``t``.
+        """
+        self.robin_bcs_a = []
+        self.robin_bcs_b = []
+        for bc in self.robin_boundaries:
+            T_inf = np.interp(t, bc.time_values, bc.values)
+            self.robin_bcs_a.append(
+                bc.h
+                * self.dT
+                * self.T_
+                * self.ds(self.dolfin_tags[self.boundary_dim][bc.boundary_name])
+            )
+            self.robin_bcs_b.append(
+                bc.h
+                * T_inf
+                * self.T_
+                * self.ds(self.dolfin_tags[self.boundary_dim][bc.boundary_name])
+            )
+
+    def update_cavern_bcs(self, cavern_handler: CavernHandler):
+        self.cavern_bcs_a = []
+        self.cavern_bcs_b = []
+        for cavern in (
+            cavern_handler.caverns_PT
+            + cavern_handler.caverns_MFlux
+            + cavern_handler.caverns_T
+        ):
+            T_inf, h = compute_cavern_heat_flux(cavern)
+            self.cavern_bcs_a.append(
+                h
+                * self.dT
+                * self.T_
+                * self.ds(self.dolfin_tags[self.boundary_dim][cavern.cavern_name])
+            )
+            self.cavern_bcs_b.append(
+                h
+                * T_inf
+                * self.T_
+                * self.ds(self.dolfin_tags[self.boundary_dim][cavern.cavern_name])
+            )
