+++ safeincave/BC/Momentum/__init__.py
@@ -2,7 +2,7 @@
 #
 # SPDX-License-Identifier: BSD-3-Clause
 
-# Backward-compatible shim; actual implementation moved to BC.Momentum
-from .BC.Momentum import GeneralBC, DirichletBC, NeumannBC, BcHandler  # noqa: F401
+from ..base import GeneralBC, DirichletBC, NeumannBC
+from .handler import BcHandler
 
 __all__ = ["GeneralBC", "DirichletBC", "NeumannBC", "BcHandler"]
