+++ safeincave/BC/__init__.py
@@ -1,0 +1,14 @@
+# Copyright (c) 2026, The SafeInCave Developers
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+from .base import GeneralBC, DirichletBC, NeumannBC
+from .cavern_utils import compute_cavern_pressure_load, compute_cavern_heat_flux
+
+__all__ = [
+    "GeneralBC",
+    "DirichletBC",
+    "NeumannBC",
+    "compute_cavern_pressure_load",
+    "compute_cavern_heat_flux",
+]
