+++ safeincave/Equations/Heat/__init__.py
@@ -1,0 +1,7 @@
+# Copyright (c) 2026, The SafeInCave Developers
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+from .base import HeatDiffusion
+
+__all__ = ["HeatDiffusion"]
