+++ safeincave/Materials/__init__.py
@@ -1,0 +1,8 @@
+# Copyright (c) 2026, The SafeInCave Developers
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+from .Material import Material
+from .Constitutive import __all__ as _CONSTITUTIVE_ALL
+
+__all__ = ["Material"] + _CONSTITUTIVE_ALL
