+++ safeincave/Mesh/__init__.py
@@ -1,0 +1,8 @@
+# Copyright (c) 2026, The SafeInCave Developers
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+from .Grid import GridHandlerGMSH
+from .MeshParameter import Element, ModelML
+
+__all__ = ["GridHandlerGMSH", "Element", "ModelML"]
