+++ safeincave/Output/SimLogging.py
@@ -597,180 +597,3 @@ class SimulationLogging:
 
         return log_kwargs
 
-
-# ============================================================================
-# Backward Compatibility Wrapper
-# ============================================================================
-
-class SimulationLoggingLegacy(SimulationLogging):
-    """
-    Legacy interface for SimulationLogging with backward compatibility.
-    
-    This class maintains the old API including auto_configure() and log_state().
-    Deprecated: Use SimulationLogging instead.
-    """
-    
-    def __init__(
-        self,
-        target_point: list | np.ndarray,
-        variables_to_track: Optional[List[str]] = None,
-    ):
-        """Initialize with legacy interface."""
-        super().__init__(target_point=target_point, variables_to_track=variables_to_track)
-    
-    def configure(
-        self,
-        centroids: np.ndarray,
-        log_file: str | None = None,
-        mom_eq=None,
-        bulk_modulus: Optional[to.Tensor] = None,
-        shear_modulus: Optional[to.Tensor] = None,
-        time_conversion: Optional[float] = None,
-    ) -> None:
-        """
-        Configure with legacy interface.
-        
-        .. deprecated::
-            Use the new interface without mom_eq parameter.
-        """
-        import warnings
-        warnings.warn(
-            "SimulationLoggingLegacy is deprecated. Use SimulationLogging instead.",
-            DeprecationWarning,
-            stacklevel=2
-        )
-        
-        # Update time_conversion if provided
-        if time_conversion is not None:
-            self.time_conversion = time_conversion
-        
-        # Call parent configure
-        super().configure(centroids, log_file)
-        
-        # Store bulk/shear modulus if provided
-        if bulk_modulus is not None:
-            self.bulk_modulus = bulk_modulus
-        if shear_modulus is not None:
-            self.shear_modulus = shear_modulus
-    
-    def auto_configure(
-        self,
-        grid,
-        material,
-        outputs: Optional[List] = None,
-        log_file: Optional[str] = None,
-        time_conversion: Optional[float] = None,
-    ) -> None:
-        """
-        Legacy auto-configure method.
-        
-        .. deprecated::
-            Use configure() directly with centroids and log_file instead.
-        """
-        import warnings
-        warnings.warn(
-            "auto_configure() is deprecated. Use configure() with centroids and log_file instead.",
-            DeprecationWarning,
-            stacklevel=2
-        )
-        
-        # Extract centroids from grid
-        centroids = self.get_centroids(grid)
-        
-        # Auto-detect log file from outputs if not provided
-        actual_log_file = log_file
-        if actual_log_file is None and outputs:
-            for output in outputs:
-                if hasattr(output, 'output_folder') and output.output_folder:
-                    import os
-                    actual_log_file = os.path.join(output.output_folder, "log.csv")
-                    break
-        
-        # Extract bulk/shear modulus from material if available
-        bulk_mod = None
-        shear_mod = None
-        if material is not None:
-            if hasattr(material, 'K_elems'):
-                bulk_mod = material.K_elems
-                shear_mod = material.G_elems
-            elif hasattr(material, 'C') and material.C is not None and material.C.shape[0] > 0:
-                C = material.C
-                bulk_mod = (C[:, 0, 0] + C[:, 1, 1] + C[:, 2, 2] + 2*(C[:, 0, 1] + C[:, 0, 2] + C[:, 1, 2])) / 9.0
-                shear_mod = C[:, 3, 3]
-        
-        # Call configure
-        self.configure(
-            centroids=centroids,
-            log_file=actual_log_file,
-            bulk_modulus=bulk_mod,
-            shear_modulus=shear_mod,
-            time_conversion=time_conversion,
-        )
-    
-    def log_state(
-        self,
-        step: int,
-        stress: to.Tensor,
-        F: to.Tensor,
-        delta_lambda: to.Tensor,
-        YieldDP: to.Tensor,
-        iteration: int = 0,
-        nonlinear_error: float = 0.0,
-        time: float = 0.0,
-        time_step: float = 0.0,
-        time_final: float = 1.0,
-        strain: Optional[to.Tensor] = None,
-        sm: Optional[to.Tensor] = None,
-        svM: Optional[to.Tensor] = None,
-    ) -> None:
-        """
-        Legacy log_state method.
-        
-        .. deprecated::
-            Use log_step() with new interface instead.
-        """
-        import warnings
-        warnings.warn(
-            "log_state() is deprecated. Use log_step() with new interface.",
-            DeprecationWarning,
-            stacklevel=2
-        )
-        
-        # Convert legacy call to new interface
-        kwargs = {
-            'yield_function': F,
-            'plastic_multiplier': delta_lambda,
-            'yield_indicator': YieldDP,
-        }
-        
-        # If sm and svM are provided, override the extractors temporarily
-        original_vars = self.variables.copy()
-        if sm is not None and 'sm' in self.variables:
-            def make_sm_extractor(sm_tensor):
-                def extractor(elem_id, **kwargs):
-                    return sm_tensor[elem_id].item()
-                return extractor
-            self.variables['sm'] = ('sm (Pa)', make_sm_extractor(sm))
-        
-        if svM is not None and 'svM' in self.variables:
-            def make_svm_extractor(svm_tensor):
-                def extractor(elem_id, **kwargs):
-                    return svm_tensor[elem_id].item()
-                return extractor
-            self.variables['svM'] = ('svM (Pa)', make_svm_extractor(svM))
-        
-        # Call new log_step
-        self.log_step(
-            step=step,
-            time=time,
-            time_step=time_step,
-            time_final=time_final,
-            iteration=iteration,
-            nonlinear_error=nonlinear_error,
-            stress=stress,
-            strain=strain,
-            **kwargs
-        )
-        
-        # Restore original variables
-        self.variables = original_vars
