+++ safeincave/Thermo/__init__.py
@@ -1,0 +1,7 @@
+# Copyright (c) 2026, The SafeInCave Developers
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+from .CavernThermodynamics import CavernThermodynamics
+
+__all__ = ["CavernThermodynamics"]
